New navigation with theme augmentation
This should replace Menu, which does not have much ability to be
styled/customized. Using Navigation allows customizing things like
the collapse open/close icons and the styling of the individual
components used inside Navigation using the theme.

Unlike in Menu, the variant must be set as a prop and is not
automatically dependent on the window size. This allows for more
control, like giving certain pages more width by always using the
temporary variant on those pages.